\name{track.ff-package}
\alias{track.ff-package}
\docType{package}
\title{
Overview of track.ff package
}
\description{

  The track.ff package glues the track package and the ff package
  together, making it easy to manage ff objects in a tracked
  database.  The ff datafiles are stored in a subdirectory of the
  tracking directory.

  This package is not needed to access ff objects in a tracking
  database, for that, only the \code{track} and \code{ff} packages are
  needed.

}
\author{Tony Plate \email{tplate@acm.org}}
\seealso{
  \link{track.ff} for description of functions and examples.
  The \link[ff:ff]{ff} package. and the \link[track:track-package]{track} package.
}
\keyword{ package }
\keyword{ data }
\keyword{ database }
\keyword{ utilities }
