\name{track.ff-package}
\alias{track.ff-package}
\docType{package}
\title{
Overview of \code{track.ff} package
}
\description{

  The \code{track.ff} package glues the track package and the \code{ff} package
  together, making it easy to manage \code{ff} objects in a tracked
  database.  The \code{ff} datafiles are stored in a subdirectory of the
  tracking directory.  The \code{track.ff} package is geared towards working
  with permanent \code{ff} objects, although it does not interfere in the
  creation of temporary \code{ff} objects.

  This package is not needed to access \code{ff} objects in a tracking
  database, for that, only the \code{track} and \code{ff} packages are
  needed.

}
\author{Tony Plate \email{tplate@acm.org}}
\seealso{
  \link{track.ff} for description of functions and examples.
  The \link[ff:ff]{ff} package. and the \link[track:track-package]{track} package.
}
\keyword{ package }
\keyword{ data }
\keyword{ database }
\keyword{ utilities }
